feat: rename raw_events into mcap_to_events
