+++ projects/owa-data/owa/data/processing/__init__.py
@@ -1,7 +1,7 @@
 """Processing module for OWA data pipeline."""
 
 from .event_to_fsl import create_fsl_dataset_from_events
-from .raw_events import create_event_dataset_from_mcaps
+from .mcap_to_events import create_event_dataset_from_mcaps
 from .resampler import create_resampler
 
 __all__ = [
